Divide toolbar into several modules
- Note: Still significantly dysfunctional
- Break apart monolithic "toolbar" module into smaller modules
  responsible for specific functions
