+++ src/modules/interface/toolbar/term-controls/append.mts
@@ -1,0 +1,132 @@
+import type { TermAbstractControl } from "/dist/modules/interface/toolbar/term-control.mjs";
+import type { ToolbarTermControlInterface } from "/dist/modules/interface/toolbar.mjs";
+import { TermInput } from "/dist/modules/interface/toolbar/term-control/term-input.mjs";
+import { TermOptionList } from "/dist/modules/interface/toolbar/term-control/term-option-list.mjs";
+import { Control } from "/dist/modules/interface/toolbar/control.mjs";
+import type { ControlFocusArea } from "/dist/modules/interface/toolbar/common.mjs";
+import { applyMatchModeToClassList, getMatchModeFromClassList } from "/dist/modules/interface/toolbar/common.mjs";
+import { type MatchMode, MatchTerm } from "/dist/modules/match-term.mjs";
+import { EleClass } from "/dist/modules/common.mjs";
+import type { SetTerm, DoPhrasesMatchTerms, ControlsInfo } from "/dist/content.mjs";
+
+class TermAppendControl implements TermAbstractControl {
+	readonly #toolbarInterface: ToolbarTermControlInterface;
+	readonly #setTerm: SetTerm;
+
+	readonly #input: TermInput;
+	readonly #optionList: TermOptionList;
+	readonly control: Control;
+	
+	readonly matchMode: MatchMode;
+
+	constructor (
+		controlsInfo: ControlsInfo,
+		toolbarInterface: ToolbarTermControlInterface,
+		setTerm: SetTerm,
+		doPhrasesMatchTerms: DoPhrasesMatchTerms,
+	) {
+		this.#toolbarInterface = toolbarInterface;
+		this.#setTerm = setTerm;
+		this.matchMode = Object.assign({}, controlsInfo.matchMode);
+		let controlContainerTemp: HTMLElement | undefined = undefined;
+		const setUpControl = (container: HTMLElement) => {
+			const pad = container.querySelector(`.${EleClass.CONTROL_PAD}`) as HTMLElement;
+			this.#input.appendTo(pad);
+			const revealButton = this.#optionList.createRevealButton();
+			revealButton.addEventListener("click", () => {
+				this.#input.classListToggle(EleClass.OPENED_MENU, true);
+			});
+			pad.appendChild(revealButton);
+			this.#optionList.appendTo(container);
+		};
+		this.control = new Control("appendTerm", {
+			buttonClasses: [ EleClass.CONTROL_BUTTON, EleClass.CONTROL_CONTENT ],
+			path: "/icons/create.svg",
+			containerId: "BAR_RIGHT",
+			setUp: container => {
+				controlContainerTemp = container;
+			},
+		}, controlsInfo, doPhrasesMatchTerms);
+		this.#input = new TermInput({ type: "append", button: this.control.button }, this, toolbarInterface);
+		this.#optionList = new TermOptionList(
+			(matchType, checked) => {
+				this.matchMode[matchType] = checked;
+				this.updateMatchModeClassList();
+			},
+			this.matchMode,
+			controlsInfo,
+			this,
+		);
+		if (controlContainerTemp) {
+			setUpControl(controlContainerTemp);
+		}
+		this.updateMatchModeClassList();
+	}
+
+	forgetToolbarOpenedMenu () {
+		this.#toolbarInterface.forgetOpenedMenu();
+	}
+
+	getInputValue () {
+		return this.#input.getValue();
+	}
+
+	selectInput (shiftCaret?: "right" | "left") {
+		this.#input.select(shiftCaret);
+	}
+
+	focusInput () {
+		return this.#input.focus();
+	}
+
+	unfocusInput () {
+		this.#input.unfocus();
+	}
+
+	openOptionList () {
+		this.#optionList.open();
+	}
+
+	getFocusArea (): ControlFocusArea {
+		if (this.#input.isFocused()) {
+			return "input";
+		}
+		return "none";
+	}
+
+	commit (inputValue?: string) {
+		inputValue ??= this.#input.getValue();
+		// TODO standard method of avoiding race condition (arising from calling termsSet, which immediately updates controls)
+		if (inputValue !== "") {
+			const matchMode: MatchMode = getMatchModeFromClassList(
+				token => this.control.classListContains(token),
+			);
+			const term = new MatchTerm(inputValue, matchMode, { allowStemOverride: true });
+			this.#setTerm(term, true);
+		}
+	}
+
+	/**
+	 * Updates the class list of the control to reflect the matching options of its term.
+	 */
+	updateMatchModeClassList () {
+		applyMatchModeToClassList(
+			this.matchMode,
+			(token, force) => this.control.classListToggle(token, force),
+		);
+	}
+
+	classListToggle (token: string, force?: boolean) {
+		return this.control.classListToggle(token, force);
+	}
+
+	classListContains (token: string) {
+		return this.control.classListContains(token);
+	}
+
+	appendTo (parent: HTMLElement) {
+		this.control.appendTo(parent);
+	}
+}
+
+export { TermAppendControl };
